refactor: Add notification controller hook, make push defaultly enabled
