+++ packages/client/components/client/Controller.ts
@@ -8,7 +8,7 @@ import { CONFIGURATION } from "@revolt/common";
 import { ModalControllerExtended } from "@revolt/modal";
 import type { State as ApplicationState } from "@revolt/state";
 import type { Session } from "@revolt/state/stores/Auth";
-import { killServiceWorkerSubscription } from "./PushNotifications";
+import { killServiceWorkerSubscription } from "./NotificationsController";
 
 export enum State {
   Ready = "Ready",
